pyproject.toml
[project]
name = "text-sentiment-analyzer"
version = "0.1.0"
description = "Simple sentiment analysis of russian economic texts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pymorphy3>=2.0.3",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]
